関数修正

+++ WindowsApp.h
@@ -25,8 +25,8 @@ private:
 
 public:
 	//ウィンドウサイズ
-	const int window_width = 1280;
-	const int window_height = 720;
+	const float window_width = 1280;
+	const float window_height = 720;
 	WNDCLASSEX w;
 	HWND hwnd;
